Fix datatype - onnx only has int64_t attributes. Update test model to validate timeout is read correctly.

+++ operators/azure/cloud_base_kernel.cc
@@ -5,6 +5,8 @@
 
 #include <sstream>
 
+#include "narrow.h"
+
 namespace ort_extensions {
 CloudBaseKernel::CloudBaseKernel(const OrtApi& api, const OrtKernelInfo& info)
     : BaseKernel(api, info),
@@ -22,7 +24,7 @@ CloudBaseKernel::CloudBaseKernel(const OrtApi& api, const OrtKernelInfo& info)
 
   model_name_ = TryToGetAttributeWithDefault<std::string>(kModelName, "");
   model_ver_ = TryToGetAttributeWithDefault<std::string>(kModelVer, "0");
-  timeout_seconds_ = TryToGetAttributeWithDefault<int>(kTimeout, kDefaultTimeoutSeconds);
+  timeout_seconds_ = narrow<int>(TryToGetAttributeWithDefault<int64_t>(kTimeout, kDefaultTimeoutSeconds));
   verbose_ = TryToGetAttributeWithDefault<std::string>(kVerbose, "0") != "0";
 
   OrtStatusPtr status{};
